[step2] refactor : id 값을 모두 Long에서 int로 변경

+++ src/main/java/com/kakao/cafe/service/UserService.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserService {
-    Long join(User user);
+    int join(User user);
 
     List<User> findUsers();
 
-    Optional<User> findOneUser(Long id);
+    Optional<User> findOneUser(int id);
 }
